CA-293886: eliminate circular imports.
